requirements.txt
python_dotenv
python-telegram-bot
supabase
google-generativeai
pydantic
pyjwt
flask
python-multipart
